# services/social-groups-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Social Groups Service API
  description: API for friend lists and guild chat
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.firemud.com
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AddFriendRequest:
      type: object
      properties:
        tenantId:
          type: string
        accountId:
          type: integer
        friendAccountId:
          type: integer
        accountLevel:
          type: boolean
      required:
        - tenantId
        - accountId
        - friendAccountId
    FriendLinkDto:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: string
        accountId:
          type: integer
        friendAccountId:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    FriendPresenceDto:
      type: object
      properties:
        friendAccountId:
          type: integer
        online:
          type: boolean
        gameInstanceId:
          type: integer
        characterId:
          type: integer
        characterName:
          type: string
        activityState:
          type: string
          enum: [ACTIVE, AUTO_AFK, EXPLICIT_AFK]
        lastSeenAt:
          type: string
          format: date-time
    SendMailRequest:
      type: object
      properties:
        tenantId:
          type: string
        senderAccountId:
          type: integer
        recipientAccountId:
          type: integer
        subject:
          type: string
        content:
          type: string
      required:
        - tenantId
        - senderAccountId
        - recipientAccountId
        - subject
        - content
    MailMessageDto:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: string
        senderAccountId:
          type: integer
        recipientAccountId:
          type: integer
        subject:
          type: string
        content:
          type: string
        sentAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
    SendMessageRequestDto:
      type: object
      properties:
        tenantId:
          type: string
        senderAccountId:
          type: integer
        type:
          type: string
          enum: [SAY, WHISPER, TELL, GUILD, CITY, ACCOUNT]
        channelId:
          type: string
        recipientAccountId:
          type: integer
        guildId:
          type: integer
        cityId:
          type: integer
        content:
          type: string
      required:
        - tenantId
        - senderAccountId
        - type
        - content
    ChatMessageDto:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: string
        senderAccountId:
          type: integer
        content:
          type: string
        timestamp:
          type: string
          format: date-time
        guildId:
          type: integer
        cityId:
          type: integer
        recipientAccountId:
          type: integer
        type:
          type: string
    AddGuildStorageItemRequest:
      type: object
      properties:
        tenantId:
          type: string
        guildId:
          type: integer
        itemName:
          type: string
        quantity:
          type: integer
      required:
        - tenantId
        - guildId
        - itemName
    GuildStorageItemDto:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: string
        guildId:
          type: integer
        itemName:
          type: string
        quantity:
          type: integer
    CreateAllianceRequest:
      type: object
      properties:
        tenantId:
          type: string
        guildId:
          type: integer
        allyGuildId:
          type: integer
      required:
        - tenantId
        - guildId
        - allyGuildId
    GuildAllianceDto:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: string
        guildId:
          type: integer
        allyGuildId:
          type: integer
        createdAt:
          type: string
          format: date-time
    CreateGuildRequest:
      type: object
      properties:
        tenantId:
          type: string
        ownerAccountId:
          type: integer
        name:
          type: string
      required:
        - tenantId
        - ownerAccountId
        - name
    GuildDto:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: string
        name:
          type: string
        ownerAccountId:
          type: integer
        createdAt:
          type: string
          format: date-time
    VoiceTokenRequestDto:
      type: object
      properties:
        tenantId:
          type: string
        accountId:
          type: integer
        channelId:
          type: string
      required:
        - tenantId
        - accountId
        - channelId
    VoiceTokenDto:
      type: object
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
    AddGuildMemberRequest:
      type: object
      properties:
        tenantId:
          type: string
        guildId:
          type: integer
        accountId:
          type: integer
        role:
          type: string
      required:
        - tenantId
        - guildId
        - accountId
        - role
    UpdateGuildMemberRoleRequest:
      type: object
      properties:
        tenantId:
          type: string
        guildId:
          type: integer
        accountId:
          type: integer
        role:
          type: string
      required:
        - tenantId
        - guildId
        - accountId
        - role
    GuildMemberDto:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: string
        guildId:
          type: integer
        accountId:
          type: integer
        role:
          type: string
paths:
  /ping:
    get:
      summary: Health check endpoint
      operationId: ping
      responses:
        '200':
          description: Pong response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: pong
        '400':
          description: Invalid request
  /friends:
    post:
      summary: Create a friend link
      operationId: addFriend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFriendRequest'
      responses:
        '200':
          description: Friend link created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/FriendLinkDto'
        '400':
          description: Invalid request
  /friends/presence:
    get:
      summary: List account-scoped friend presence
      operationId: listFriendPresence
      parameters:
        - in: query
          name: tenantId
          required: true
          schema:
            type: integer
        - in: query
          name: accountId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Friend presence list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FriendPresenceDto'
        '400':
          description: Invalid request
  /mail:
    post:
      summary: Send in-game mail
      operationId: sendMail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMailRequest'
      responses:
        '200':
          description: Mail sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/MailMessageDto'
        '400':
          description: Invalid request
  /chat:
    post:
      summary: Send chat message
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequestDto'
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ChatMessageDto'
        '400':
          description: Invalid request
  /guilds:
    post:
      summary: Create guild
      operationId: createGuild
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGuildRequest'
      responses:
        '200':
          description: Guild created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GuildDto'
        '400':
          description: Invalid request
  /guilds/storage:
    post:
      summary: Add item to guild storage
      operationId: addStorageItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGuildStorageItemRequest'
      responses:
        '200':
          description: Item added
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GuildStorageItemDto'
        '400':
          description: Invalid request
  /guilds/alliances:
    post:
      summary: Create guild alliance
      operationId: createAlliance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAllianceRequest'
      responses:
        '200':
          description: Alliance created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GuildAllianceDto'
        '400':
          description: Invalid request
  /guilds/members:
    post:
      summary: Add guild member
      operationId: addGuildMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGuildMemberRequest'
      responses:
        '200':
          description: Member added
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GuildMemberDto'
        '400':
          description: Invalid request
  /guilds/members/role:
    post:
      summary: Update guild member role
      operationId: updateGuildMemberRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGuildMemberRoleRequest'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GuildMemberDto'
        '400':
          description: Invalid request
  /guilds/members/remove:
    post:
      summary: Remove guild member
      operationId: removeGuildMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGuildMemberRequest'
      responses:
        '200':
          description: Member removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
        '400':
          description: Invalid request
  /voice/token:
    post:
      summary: Issue voice chat token
      operationId: createVoiceToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoiceTokenRequestDto'
      responses:
        '200':
          description: Voice token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/VoiceTokenDto'
        '400':
          description: Invalid request